Wrote the algorithm

+++ src/algorithms/heuristics.h
@@ -1,5 +1,5 @@
 #include "../structures/tile.h"
 #ifndef HEURISTICS_H
 #define HEURISTICS_H
-float cartesian(tile* current,tile* goal);
+double cartesian(tile* current,tile* goal);
 #endif
